Create response.h
I tested processResponse and it was working for. Needs the buffer to be passed to it and the type to be passed to it (GPS, etc)
The formatting or protocol that it uses is [response: 1]/r/n[type]/r/n[length]/r/n[data]

Haven't tested sendResponse but I am pretty sure it works. If not it should be an easy fix.
Needs the buffer, the udp client socket, and the client address.